misc casts in psa

+++ library/psa_crypto_cipher.c
@@ -349,9 +349,9 @@ static psa_status_t psa_cipher_setup(
 #endif /* MBEDTLS_PSA_BUILTIN_ALG_CBC_NO_PADDING ||
           MBEDTLS_PSA_BUILTIN_ALG_CBC_PKCS7 */
 
-    operation->block_length = (PSA_ALG_IS_STREAM_CIPHER(alg) ? 1 :
-                               PSA_BLOCK_CIPHER_BLOCK_LENGTH(key_type));
-    operation->iv_length = PSA_CIPHER_IV_LENGTH(key_type, alg);
+    operation->block_length = (uint8_t) (PSA_ALG_IS_STREAM_CIPHER(alg) ? 1 :
+                                         PSA_BLOCK_CIPHER_BLOCK_LENGTH(key_type));
+    operation->iv_length = (uint8_t) PSA_CIPHER_IV_LENGTH(key_type, alg);
 
 exit:
     return mbedtls_to_psa_error(ret);
